requirements.txt
torchvision==0.5.0
tensorboard~=1.15.0
Pillow<7.0.0
tqdm==4.42.1
scikit-learn==0.22.1
pandas~=0.25.3
matplotlib~=3.1.3